Shut up bot (#27)
* Allow hiding only useless bot messages.
This is enabled by adding a Useful class to useful bot responses.
Setting it causes subsequent messages to not have their useless flag
set to true. (It can be passed with a false parameter if some messages
in the list are useful and some are useless.) All bot messages from
the codex script except for the response to adding a quip is considered
to be useful. All other bot commands are assumed to be useless.
I could be convinced that the "hooray" after setting an answer should be
marked useless. I could also be convinced that error messages are useless,
but someone who might want to hide useless bot text might still want to see
them, especially if they were typing bot commands.

* Hide commands that trigger useless bot responses.
Doesn't hide your own commands because that would be confusing.