Enable desktop notifications from Oplog messages (#32)
* Add stream field to oplogs for notification deduping/muting.
Also made it part of share.model and set the field on existing calls.

* Add "bot announce" command to add to oplog.

* Add answer to oplog for setAnswer.
Previously it was only in the created/canceled callins, but with
notifications, someone might subscribe to answers but not callins.

* Use JS Notification API to add desktop notifications for oplog events.

* Add UI to modify which notifications are listened for.
Also changes renotify and tag, as the intent was to both dedupe and
rate-limit, but at least on mac, a second notification for the same
stream won't appear if you didn't manually close the first, which is
probably hidden in the notifications menu you'd never check.