Use Meteor's native accounts system (#65)
Because our old setup used HTTP Basic Authentication on the initial page load but not on the websocket the app uses to download its data (because websockets and basic auth don't mix), I believe the app had a security vulnerability where someone could either compile a Cordova version of the app from the (open-)source, sideload it onto a phone, and connect to the app unauthenticated, or just open a websocket connection themselves and talk DDS to the backend. This replaces that auth scheme with a primitive custom Meteor accounts implementation: There's no account creation per se, and only a shared team password. If you have it, you can log in, which creates your nickname if it didn't already exist. This also means we could run more easily on a PaaS like Galaxy.
This also replaces all 'who' parameters with the userid of the connection, so you can't impersonate someone (though that would only matter if there were different privilege levels).