Replace reactive-storage with our own implementation
It seems to have a bug where it gets into an infinite loop if you set a value twice in quick succession and have three windows open, because it writes the values back to the storage area when it may already have been updated with the replacement value. This version just keeps track of the dependencies.