Support any number of metapuzzles per puzzle (#74)
Schema change that allows a puzzle to feed into any number of metapuzzles. The object types are now:

* Puzzle: anything that has an answer
* Round: a webpage with puzzles on it.

Old Round Groups are now Rounds; old Rounds are now included in Puzzles.

Metapuzzles are distinguished from 'leaf' puzzles by the presence of a 'puzzles' field containing a list; a puzzle can be marked as a meta without any puzzles feeding it. Also, all puzzles now have a feedsInto field containing a list of metapuzzles they feed into. MongoDB makes both normalized and denormalized forms a pain, but denormalized is comparatively less.

New methods to support the schema change are:

* makeMeta
* makeNotMeta
* feedMeta
* unfeedMeta

The bot supports equivalent commands.

Visually, metas are rendered under the Round that contains them. Non-meta puzzles are rendered in each meta that contains them, which doesn't need to be the same round. If a puzzle feeds no metas, it is rendered in an "uncategorized" section in its round. This applies both to the main table and the status grid. In edit mode, a dropdown is provided in the tag editing section of each puzzle containing all metas it doesn't currently feed to allow it to be marked as feeding them. Metas it currently feeds are listed with a delete button to unfeed them. Puzzles can be marked as metas with a checkbox, and metas with no puzzles feeding them can become leaf puzzles by unchecking the box.

In the breadcrumbs, BFS is used to find all the ancestor metas of a puzzle. Each is listed in the breadcrumbs at the closest path to the puzzle. If there are multiple metas at that distance, they are collapsed into a dropdown. Cycles, while unlikely in a real hunt, aren't a problem because BFS is used.