Don't send deleted messages to the client
Follow-up to #123.
Because the last read message marker was based on the exact timestamp of the message it was below, if you deleted a message but didn't keep sending it (or at least am empty stub) to the client, anyone who had read exactly to that message would have their read marker disappear. This change uses a MutationObserver to position the read marker between the last message with data-read="read" and the first with data-read="unread". To avoid recalculating the read attribute unnecessarily, we use a nonreactive read of the timestamp at first, and only register a dependency on it if the message is unread.