Keep callins (#162)
* Leave resolved callins in the CallIns table

Adds a new status field that starts at 'pending' and can change to 'accepted', 'rejected', or 'cancelled'. This will help with #149 (and will replace the incorrectAnswers field on Puzzles), and pending callins for #147 can go in the same table.

* List pending and completed callins in puzzle info

Instead of copying incorrect answers to a special list, just list all callins for a puzzle on its info along with resolution. Also allows resolving or cancelling a callin directly from the puzzle page.

* Attach the caller-in to pending callins

* Update tests to not look for incorrectAnswers

Since we're leaving resolved answers in the callins table, we don't need to add them to an incorrectanswers array, so we won't. The tests still thought we would.

* Put resolution buttons back in table cell.