Keepalive in chat autorun, not onCreated
More pages have chat now, which means it's possible to go directly from a page with chat to another. This means we need to join the new room when the room_name changes, not in a `messages.onCreated` callback, since the callback won't be called again if the chat room changes to another.
Fixes #156