Docs (#234)
* Copy wiki content to docs directory

* Update links to reflect de-wikification

* Team leads may override the use of the queue.

* Add some line breaks

* More updates

* Clarify that notification will happen

* Remove quips path

Don't make people use it. The bot is better.

* Docs (#233)

* Update Solving.md

Rewrite of Jitsi Meetings and Tags section. Minor formatting fixes elsewhere.

* Update Solving.md

Bold that you start meeting muted and no video.

* Solver.md Answer section cleanup

Moving all of the user - how to do X before the long explanation of why we do it this way.

Did I cut too much of the explanation about why/how to use each type of call-in?

* solving.md - make starring messages in chat more prominent.

* Solving.md - typo

Co-authored-by: ThornInJonsSide <56458313+ThornInJonsSide@users.noreply.github.com>

* Fix mechanics link typo

Co-authored-by: ThornInJonsSide <56458313+ThornInJonsSide@users.noreply.github.com>