Remove compatibility shims (#267)
Any callin created with the 2021 version has a status, so we don't need to set it. Similarly, it doesn't have the target + answer index so we don't need to drop it.
Users won't have their gravatar set with a stored email address, so we don't need to convert it to an md5.